Desloppify code
